update: extract to declaration file next-auth
